requirements_test.txt
atomicwrites==1.4.0
attrs==20.1.0
certifi==2020.6.20
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
iniconfig==1.0.1
jsons==1.2.0
kiwisolver==1.2.0
matplotlib==3.3.1
more-itertools==8.5.0
networkx==2.4
numpy==1.19.1
packaging==20.4
Pillow==9.0.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.0
pyshp==2.1.0
pytest==6.1.2
pytest-qt==3.3.0
python-dateutil==2.8.1
scipy==1.5.2
six==1.15.0
toml==0.10.1
typish==1.7.0
wcwidth==0.2.5
